Refactor telemetry service boundaries
