+++ controller/include/telemetry/telemetry_frame_normalizer.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "naim/runtime/runtime_status.h"
+
+namespace naim::controller {
+
+class TelemetryFrameNormalizer final {
+ public:
+  naim::HostTelemetryFrame Normalize(naim::HostTelemetryFrame frame) const;
+};
+
+}  // namespace naim::controller
